requirements.txt
# X-Ray SDK and API Dependencies

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.3.0

# HTTP Client
httpx==0.25.1

# Utilities
python-dotenv==1.0.0

# UI
streamlit==1.28.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
